ia-reclamation/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
torch==2.3.0
torchvision==0.18.0
pillow==10.4.0
pillow-avif-plugin==1.4.3
pandas==2.2.2
python-multipart==0.0.9